Line 클래스 생성, 필드 map의 제네릭을 Line으로 변경
