+++ src/main/java/ladder/domain/Line.java
@@ -1,0 +1,13 @@
+package ladder.domain;
+
+public class Line {
+    private int coordinate;
+
+    public Line(int coordinate) {
+        if (coordinate < 0) {
+            throw new IllegalArgumentException("좌표의 값은 0보다 작을 수 없습니다.");
+        }
+        this.coordinate = coordinate;
+    }
+
+}
