+++ src/test/java/PersonTest.java
@@ -1,3 +1,4 @@
+import ladder.domain.Person;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
